Allow splitting corpora configuration into separate json files.
As otherwise, the main config gets quite unreadable with more corpora.
